+++ app/src/main/java/ie/wit/models/MovieStore.kt
@@ -2,8 +2,8 @@ package ie.wit.models
 
 interface MovieStore {
     fun findAll() : ArrayList<MovieModel>
-    fun findById(id: Long) : MovieModel?
+    fun findById(id: Int) : MovieModel?
     fun create(movie: MovieModel)
     fun update(movie: MovieModel)
-    fun deleteMovie(id: Long)
+    fun deleteMovie(id: Int)
 }
